style: added icons for sidebar and seperate the files
